requirements.txt
torch
transformers>=4.36.2
datasets
numpy
tqdm
PyYAML
